+++ build.config.ts
@@ -1,6 +1,0 @@
-import { defineBuildConfig } from 'unbuild';
-
-export default defineBuildConfig({
-  clean: true,
-  externals: ['zustand', 'zustand/middleware', 'rxjs', 'immer', 'react'],
-});